+++ Changelog.md
@@ -3,6 +3,7 @@
 ## v1.3.1a
 * Added ap_utils.makeRocIfNameContains
 * Added overflow paging for comms menus
+* Added unit_repairman helper to schedule periodic respawns of named units
 ## v1.3a
 * Combined all scripts into one file
 * Fixed ap_utils.generateGroups issue with airbourne spawns